Move UndefinedVariable into CleanCode namespace
