+++ src/test/resources/files/Rule/CleanCode/UndefinedVariable/testRuleAppliesToUndefinedVariable.php
