+++ src/test/resources/files/Rule/CleanCode/UndefinedVariable/testRuleAppliesToUndefinedVariableWithDefinedVariable.php
