+++ src/test/resources/files/Rule/CleanCode/UndefinedVariable/testRuleDoesNotApplyToSuperGlobals.php
