+++ src/test/resources/files/Rule/CleanCode/UndefinedVariable/testRuleDoesNotApplyToUsedProperties.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * This file is part of PHP Mess Detector.
+ *
+ * Copyright (c) Manuel Pichler <mapi@phpmd.org>.
+ * All rights reserved.
+ *
+ * Licensed under BSD License
+ * For full copyright and license information, please see the LICENSE file.
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @author Manuel Pichler <mapi@phpmd.org>
+ * @copyright Manuel Pichler. All rights reserved.
+ * @license https://opensource.org/licenses/bsd-license.php BSD License
+ * @link http://phpmd.org/
+ */
+
+class testRuleDoesNotApplyToUsedProperties extends AbstractTest
+{
+    protected $x = 'abc';
+
+    function testRuleDoesNotApplyToUsedProperties()
+    {
+        echo $this->x;
+    }
+}
